Moved noise-pass to noise
I am unsure where this file should go, it will stay in noise for now
(The candidates are noise and root)
