--- src/noise-gen/noise-pass.ts
@@ -1,3 +1,0 @@
-export { getFBM } from './noise/fBm'
-export { GradientNoise } from './noise/gradient-noise'
-export { ValueNoise } from './noise/value-noise'
