+++ src/noise-gen/noise/noise-pass.ts
@@ -1,0 +1,3 @@
+export { getFBM } from './fBm'
+export { GradientNoise } from './gradient-noise'
+export { ValueNoise } from './value-noise'
